% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{check_mnis_id}
\alias{check_mnis_id}
\title{Check mnis id}
\usage{
check_mnis_id(member_mnis_id, house)
}
\arguments{
\item{member_mnis_id}{The mnis id of a member to check.}

\item{house}{Commons or Lords}
}
\description{
Check mnis id
}
\keyword{internal}
