% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clvotes.R
\docType{package}
\name{clvotes}
\alias{clvotes}
\title{clvotes: Tools for retrieving data from the Commons and Lords
Votes API}
\description{
The clvotes package provides a suit of tools for downloading and
processing data from the UK Parliament's Commons and Lords Votes API.
}
