% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_commons.R
\name{fetch_cds_raw}
\alias{fetch_cds_raw}
\title{Fetch data on all Commons divisions available from the API}
\usage{
fetch_cds_raw()
}
\description{
Fetch data on all Commons divisions available from the API
}
\keyword{internal}
