% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_commons.R
\name{fetch_cds_raw_filter}
\alias{fetch_cds_raw_filter}
\title{Filter data on all Commons divisions}
\usage{
fetch_cds_raw_filter(from_date = NA, to_date = NA, on_date = NA)
}
\description{
Filter data on all Commons divisions
}
\keyword{internal}
