% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_commons.R
\name{fetch_cds_votes_raw}
\alias{fetch_cds_votes_raw}
\title{Fetch data on individual Commons divisions available from the API}
\usage{
fetch_cds_votes_raw(division_id)
}
\description{
Fetch data on individual Commons divisions available from the API
}
\keyword{internal}
