% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_raw.R
\name{fetch_commons_divisions_all_raw}
\alias{fetch_commons_divisions_all_raw}
\title{Fetch key details on all Commons divisions in raw format}
\usage{
fetch_commons_divisions_all_raw(from_date = NA, to_date = NA, on_date = NA)
}
\arguments{
\item{from_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the from_date.}

\item{to_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the to_date.}

\item{on_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the on_date.}
}
\value{
A tibble of key details for each Commons division, with one row
per division in raw format.
}
\description{
\code{fetch_commons_divisions_all_raw} fetches a dataframe from the Commons Votes API
showing key details about each division, with one row per division in raw
format.
}
\details{
JSON returned in the response from the API is converted into a Tibble with
column names cleaned. No sub-selecting of columns and variable names occurs.

The from_date and to_date arguments can be used to filter divisions based
on the dates they occurred. The on_date argument is a convenience that sets
the from_date and to_date to the same given date. The on_date has priority:
if the on_date is set, the from_date and to_date are ignored.
}
