% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_commons.R
\name{fetch_commons_divisions_party}
\alias{fetch_commons_divisions_party}
\title{Fetch key details on Commons divisions votes grouped by party}
\usage{
fetch_commons_divisions_party(division_id = NULL)
}
\arguments{
\item{division_id}{An integer or vector of integers representing a division
ID.}
}
\value{
A tibble of key details of voting grouped by party for a Commons
division, with one row per grouped vote.
}
\description{
\code{fetch_commons_divisions_party} fetches data from the Commons Votes API
showing key details about each division vote grouped by party, with one row
per division grouped vote. All votes or abstentions are returned.
}
