% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_commons.R
\name{fetch_commons_divisions_votes}
\alias{fetch_commons_divisions_votes}
\title{Fetch key details on Commons divisions votes}
\usage{
fetch_commons_divisions_votes(division_id)
}
\arguments{
\item{division_id}{An integer or vector of integers representing a division
ID.}
}
\value{
A tibble of key details of voting for a Commons division, with one row
per vote.
}
\description{
\code{fetch_commons_divisions_votes} fetches data from the Commons Votes API
showing key details about each division vote, with one row per division
vote.
}
