% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_lords.R
\name{fetch_lds_members_raw}
\alias{fetch_lds_members_raw}
\title{Fetch data on Lords voting for divisions available from the API}
\usage{
fetch_lds_members_raw(
  member_mnis_id,
  from_date = NULL,
  to_date = NULL,
  on_date = NULL
)
}
\description{
Fetch data on Lords voting for divisions available from the API
}
\keyword{internal}
