% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_lords.R
\name{fetch_lds_raw}
\alias{fetch_lds_raw}
\title{Fetch data on all Lords divisions available from the API}
\usage{
fetch_lds_raw()
}
\description{
Fetch data on all Lords divisions available from the API
}
\keyword{internal}
