% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_lords.R
\name{fetch_lds_raw_filter}
\alias{fetch_lds_raw_filter}
\title{Filter data on all Lords divisions}
\usage{
fetch_lds_raw_filter(from_date = NA, to_date = NA, on_date = NA)
}
\description{
Filter data on all Lords divisions
}
\keyword{internal}
