% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_lords.R
\name{fetch_lds_votes_raw}
\alias{fetch_lds_votes_raw}
\title{Fetch data on individual Lords divisions available from the API}
\usage{
fetch_lds_votes_raw(division_id)
}
\description{
Fetch data on individual Lords divisions available from the API
}
\keyword{internal}
