% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_lords.R
\name{fetch_lords_divisions_lord_type}
\alias{fetch_lords_divisions_lord_type}
\title{Fetch key details on Lords divisions votes grouped by lord type}
\usage{
fetch_lords_divisions_lord_type(division_id)
}
\arguments{
\item{division_id}{An integer or vector of integers representing a division
ID.}
}
\value{
A tibble of key details of voting grouped by lord type for a Lords
division, with one row per grouped vote.
}
\description{
\code{fetch_lords_divisions_lord_type} fetches data from the Lords Votes API
showing key details about each division vote grouped by lord type, with one
row per division grouped vote.
}
