% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_lords.R
\name{fetch_lords_divisions_members}
\alias{fetch_lords_divisions_members}
\title{Fetch key details Lords voting record for divisions}
\usage{
fetch_lords_divisions_members(
  member_mnis_id,
  from_date = NA,
  to_date = NA,
  on_date = NA
)
}
\arguments{
\item{member_mnis_id}{An integer or vector of integers representing a member
mnis ID.}

\item{from_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the from_date.}

\item{to_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the to_date.}

\item{on_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NA, which means no records are excluded on the basis of
the on_date.}
}
\value{
A tibble of key details for a Lords voting record, with one row
per division.
}
\description{
\code{fetch_lords_divisions_members} fetches data from the Lords Votes API
showing key details about the voting record for a Lord for each division,
with one row per division. Only whether a Lord voted "Content" in a division
is returned.
}
\details{
The from_date and to_date arguments can be used to filter divisions based
on the dates they occurred. The on_date argument is a convenience that sets
the from_date and to_date to the same given date. The on_date has priority:
if the on_date is set, the from_date and to_date are ignored.
}
