% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_lords.R
\name{fetch_lords_divisions_party}
\alias{fetch_lords_divisions_party}
\title{Fetch key details on Lords divisions votes grouped by party}
\usage{
fetch_lords_divisions_party(division_id)
}
\arguments{
\item{division_id}{An integer or vector of integers representing a division
ID.}
}
\value{
A tibble of key details of voting grouped by party for a Lords
division, with one row per grouped vote.
}
\description{
\code{fetch_lords_divisions_party} fetches data from the Lords Votes API
showing key details about each division vote grouped by party, with one row
per division grouped vote.
}
