% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_lords.R
\name{fetch_lords_divisions_remote}
\alias{fetch_lords_divisions_remote}
\title{Fetch key details on all Lords remote divisions}
\usage{
fetch_lords_divisions_remote(from_date = NULL, to_date = NULL, on_date = NULL)
}
\arguments{
\item{from_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NULL, which means no records are excluded on the basis of
the from_date.}

\item{to_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NULL, which means no records are excluded on the basis of
the to_date.}

\item{on_date}{A string or Date representing a date. If a string is used
it should specify the date in ISO 8601 date format e.g. '2000-12-31'. The
default value is NULL, which means no records are excluded on the basis of
the on_date.}
}
\value{
A tibble of key details of remote Lords divisions, with one row
per remote division.
}
\description{
\code{fetch_lords_divisions_remote} fetches a dataframe from the Lords
Votes API showing key details about each remote division, with one row per
remote division.
}
\details{
The from_date and to_date arguments can be used to filter divisions based
on the dates they occurred. The on_date argument is a convenience that sets
the from_date and to_date to the same given date. The on_date has priority:
if the on_date is set, the from_date and to_date are ignored.
}
