% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_cds_members_raw}
\alias{format_cds_members_raw}
\title{Format raw Commons divisions members tibble}
\usage{
format_cds_members_raw(df)
}
\arguments{
\item{df}{The divisions member tibble retrieved from the API}
}
\description{
Format raw Commons divisions members tibble
}
\keyword{internal}
