% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_cds_raw}
\alias{format_cds_raw}
\title{Format raw Commons divisions tibble}
\usage{
format_cds_raw(df)
}
\arguments{
\item{df}{The divisions tibble retrieved from the API}
}
\description{
Format raw Commons divisions tibble
}
\keyword{internal}
