% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_lds_members_raw}
\alias{format_lds_members_raw}
\title{Format raw Lords divisions members tibble}
\usage{
format_lds_members_raw(df)
}
\arguments{
\item{df}{The divisions member tibble retrieved from the API}
}
\description{
Format raw Lords divisions members tibble
}
\keyword{internal}
