% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_lds_raw}
\alias{format_lds_raw}
\title{Format raw Lords divisions tibble}
\usage{
format_lds_raw(df)
}
\arguments{
\item{df}{The divisions tibble retrieved from the API}
}
\description{
Format raw Lords divisions tibble
}
\keyword{internal}
