% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_lds_votes_raw}
\alias{format_lds_votes_raw}
\title{Format raw Lords divisions votes tibble}
\usage{
format_lds_votes_raw(df)
}
\arguments{
\item{df}{The divisions vote tibble retrieved from the API}
}
\description{
Format raw Lords divisions votes tibble
}
\keyword{internal}
