% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{process_lds_pagination}
\alias{process_lds_pagination}
\title{Processes hidden pagination of results retrieved from the API}
\usage{
process_lds_pagination(url)
}
\arguments{
\item{url}{The full API URL specifying the endpoint and request parameters.}
}
\description{
Processes hidden pagination of results retrieved from the API
}
\keyword{internal}
