% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{request_error}
\alias{request_error}
\title{Report an error with a http response}
\usage{
request_error(response)
}
\arguments{
\item{response}{The text of the server response}
}
\description{
Report an error with a http response
}
\keyword{internal}
